+++ src/dynamics/diagnostic_variables.jl
@@ -171,6 +171,7 @@ $(TYPEDFIELDS)"""
     SpectralVariable3D,     # <: LowerTriangularArray
     GridVariable2D,         # <: AbstractField
     GridVariable3D,         # <: AbstractField
+    ScratchMemoryType,      # <: ScratchMemory{NF, ArrayType{Complex{NF},3}}
 } <: AbstractDiagnosticVariables
     
     spectrum::SpectrumType            # spectral resolution: maximum degree and order of spherical harmonics
@@ -236,19 +237,34 @@ $(TYPEDFIELDS)"""
 
     "Vertical average of divergence [1/s], spectral"
     div_mean::SpectralVariable2D = zeros(SpectralVariable2D, spectrum)
+
+    "Scratch memory for the transforms"
+    scratch_memory::ScratchMemoryType = SpeedyTransforms.ScratchMemory(NF, ArrayType, grid, nlayers)
 end
 
 """$(TYPEDSIGNATURES)
-Generator function."""
-function DynamicsVariables(SG::SpectralGrid)
+Generator function. If a `spectral_transform` is handed over, the same scratch memory is used."""
+function DynamicsVariables(SG::SpectralGrid; 
+                           spectral_transform::Union{Nothing,SpectralTransform}=nothing) 
     (; spectrum, grid, nlayers, NF, ArrayType) = SG
     (; SpectralVariable2D, SpectralVariable3D) = SG
     (; GridVariable2D, GridVariable3D) = SG
 
-    return DynamicsVariables{NF, ArrayType, typeof(spectrum), typeof(grid),
-        SpectralVariable2D, SpectralVariable3D, GridVariable2D, GridVariable3D}(;
-            spectrum, grid, nlayers,
-        )
+    if isnothing(spectral_transform)
+        return DynamicsVariables{NF, ArrayType, typeof(spectrum), typeof(grid),
+            SpectralVariable2D, SpectralVariable3D,
+            GridVariable2D, GridVariable3D, SpeedyTransforms.ScratchMemory{NF, ArrayType{Complex{NF}, 3}}}(;
+                spectrum, grid, nlayers,
+            )
+    else 
+        scratch_memory = spectral_transform.scratch_memory 
+
+        return DynamicsVariables{NF, ArrayType, typeof(spectrum), typeof(grid),
+            SpectralVariable2D, SpectralVariable3D,
+            GridVariable2D, GridVariable3D, typeof(scratch_memory)}(;
+                spectrum, grid, nlayers, scratch_memory
+            )
+    end 
 end
 
 
@@ -432,6 +448,7 @@ struct DiagnosticVariables{
     ParticleVector,         # <: AbstractField
     VectorType,             # <: AbstractVector
     MatrixType,             # <: AbstractMatrix
+    ScratchMemoryType,      # <: ArrayType{Complex{NF}, 3}
     Interpolator,           # <:AbstractInterpolator
 } <: AbstractDiagnosticVariables
 
@@ -455,7 +472,7 @@ struct DiagnosticVariables{
     grid::GridVariables{NF, ArrayType, GridType, GridVariable2D, GridVariable3D}
     
     "Intermediate variables for the dynamical core"
-    dynamics::DynamicsVariables{NF, ArrayType, SpectrumType, GridType, SpectralVariable2D, SpectralVariable3D, GridVariable2D, GridVariable3D}
+    dynamics::DynamicsVariables{NF, ArrayType, SpectrumType, GridType, SpectralVariable2D, SpectralVariable3D, GridVariable2D, GridVariable3D, ScratchMemoryType}
     
     "Global fields returned from physics parameterizations"
     physics::PhysicsVariables{NF, ArrayType, GridType, GridVariable2D}
@@ -474,7 +491,7 @@ struct DiagnosticVariables{
 end
 
 function DiagnosticVariables(SG::SpectralGrid, model::Union{Barotropic, ShallowWater})
-    diagn = DiagnosticVariables(SG)
+    diagn = DiagnosticVariables(SG; spectral_transform=model.spectral_transform)
     add!(diagn, model.tracers)
     return diagn
 end
@@ -483,15 +500,16 @@ end
 function DiagnosticVariables(SG::SpectralGrid, model::PrimitiveEquation)
     nbands_shortwave = get_nbands(model.shortwave_radiation)
     nbands_longwave = get_nbands(model.longwave_radiation)
-    diagn =  DiagnosticVariables(SG; nbands_shortwave, nbands_longwave)
+    diagn =  DiagnosticVariables(SG; spectral_transform=model.spectral_transform, nbands_shortwave, nbands_longwave)
     add!(diagn, model.tracers)
     return diagn
 end
 
 """$(TYPEDSIGNATURES)
-Generator function."""
+Generator function. If a `transform` is handed over, the same scratch memory is used."""
 function DiagnosticVariables(
     SG::SpectralGrid;
+    spectral_transform::Union{Nothing, SpectralTransform} = nothing,
     nbands_shortwave::Integer = 0,
     nbands_longwave::Integer = 0,
 )
@@ -499,7 +517,7 @@ function DiagnosticVariables(
 
     tendencies = Tendencies(SG)
     grid = GridVariables(SG)
-    dynamics = DynamicsVariables(SG)
+    dynamics = DynamicsVariables(SG; spectral_transform)
     physics = PhysicsVariables(SG)
     particles = ParticleVariables(SG)
     column = ColumnVariables(SG; nbands_shortwave, nbands_longwave)