+++ test/GPU/runtests.jl
@@ -12,3 +12,4 @@ include("kernels_GPU.jl")
 
 # test if the models run on GPU 
 include("barotropic.jl")
+